acm_web_backend/acm_website_core/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.20.40
botocore==1.23.40
certifi==2017.11.5
cffi==1.15.0
chardet==3.0.4
cryptography==36.0.1
distlib==0.3.1
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.13.0
django-countries==7.3.2
django-crispy-forms==1.13.0
django-environ==0.8.1
django-filter==21.1
django-states==1.6.5
django-storages==1.12.3
djangorestframework==3.12.2
environ==1.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.6
jmespath==0.10.0
Markdown==3.3.3
mysql-client==0.0.1
mysqlclient==2.1.0
paramiko==2.9.2
pbr==5.5.1
Pillow==9.0.0
psycopg2==2.9.2
psycopg2-binary==2.9.3
pycparser==2.21
pygame==2.0.0
PyMySQL==1.0.2
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2017.3
requests==2.18.4
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
stevedore==3.3.0
storage==0.0.4.3
typing_extensions==4.3.0
tzdata==2021.5
urllib3==1.21.1
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
Wrapper==1.1.0b1
xlwt==1.3.0
